requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.104.0
fastapi-cors==0.0.6
flower==2.0.1
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
humanize==4.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
orjson==3.9.9
packaging==23.2
passlib==1.7.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
vine==5.0.0
watchfiles==0.21.0
wcwidth==0.2.8
websockets==11.0.3
